[HMS ML KIT] UI Hierarchy refactor
