+++ app/src/main/res/layout/activity_mlkit_text_recognition.xml
@@ -1,0 +1,33 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.appcompat.widget.LinearLayoutCompat xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical"
+    android:padding="16dp"
+    tools:context=".ui.ml_kit.TextRecognitionActivity">
+
+    <androidx.appcompat.widget.AppCompatImageView
+        android:layout_width="match_parent"
+        android:layout_height="200dp"
+        android:src="@drawable/text_recognition" />
+
+    <com.google.android.material.button.MaterialButton
+        android:id="@+id/btn_text_recognition"
+        style="@style/ButtonStyle"
+        android:text="TEXT RECOGNITION FROM IMAGE" />
+
+    <com.google.android.material.button.MaterialButton
+        android:id="@+id/btn_live_recognition"
+        style="@style/ButtonStyle"
+        android:text="LIVE STREAM RECOGNITION" />
+
+
+    <com.cengiztoru.hmscorekits.ui.custom_view.LoggerTextView
+        android:id="@+id/tv_logger"
+        android:layout_width="match_parent"
+        android:layout_height="0dp"
+        android:layout_marginTop="16dp"
+        android:layout_weight="1"
+        android:gravity="center" />
+</androidx.appcompat.widget.LinearLayoutCompat>
